+++ TabletopTweaks-Base/NewContent/Templates/AlignmentTemplates.cs
@@ -26,8 +26,8 @@ using System.Collections.Generic;
 using System.Text.RegularExpressions;
 using TabletopTweaks.Core.Utilities;
 using TabletopTweaks.Core.Wrappers;
-using static TabletopTweaks.Core.MechanicsChanges.AdditionalModifierDescriptors;
 using static TabletopTweaks.Base.Main;
+using static TabletopTweaks.Core.MechanicsChanges.AdditionalModifierDescriptors;
 
 namespace TabletopTweaks.Base.NewContent.Templates {
     static class AlignmentTemplates {